Added muon-only datacards for Delphes simulation

+++ darkPhotonForTrigger/pythiaFragmentLeptonJetMass1_0.py
@@ -9,7 +9,7 @@ generator = cms.EDFilter("Pythia8GeneratorFilter",
     pythiaPylistVerbosity = cms.untracked.int32(1),
     filterEfficiency = cms.untracked.double(1.0),
     pythiaHepMCVerbosity = cms.untracked.bool(False),
-    comEnergy = cms.double(13000),
+    comEnergy = cms.double(14000),
     PythiaParameters = cms.PSet(
         pythia8CommonSettingsBlock,
         pythia8CP5SettingsBlock,
